refactor: split caddy.go (#1629)
* Splits modules.

* trigger build

---------
